initial disassembler. not complete yet.
currently there 2 unsupported "fixup" (relocation) types:
FIXUP_DATADATA and FIXUP_STACK.
i'm trying to evaluate when the compiler emits these as they seem
to be very rare or even unused in the STACK case.
"duty and beyond" uses a DATADATA relocation in the globalscript.

also it is unclear how to deal with globaldata, the game sees it
as a chunk of mem with offsets into it. however if we treat it
as such, endian issues appear. thus it seems to be preferable
to determine the sizes of the variables and name them.

in order to detect sizes of exported variables, we have to scan
all scripts that could use them and inspect the assembly that
accesses it.

the last missing point is to figure out what "sections" do.

once these issues are fixed, it should be possible to recompile
the assembly (with or without changes) into a valid ags-objectfile,
which can then be injected into the original datafile/roomfiles,
by storing the original start and end offsets.