working assembler
i was able to recompile a simple room script and inject it into the
original testgame, and it seemed to work.
any bigger script will not work yet, there's a big FIXME in
Assembler.c, we need something like a read_argument function that
can parse any type of argument passed in a "li" instruction,
for example escaped strings which could contain whitespace.
also, imports originally seem to have many empty entries.
it is possible that the engine has a global import list
and function indices must match in all scripts.
this would be a pain to implement, so hopefully its not necessary.

also FIXUP_GLOBALDATA needs to implemented, this should be quite
easy though. (my test room script didnt have globaldata so far.)