Script: add debug mode, and some assertions & label lookups in-mem
since we already have the code read into memory, use that rather
than to process the file again.

debugmode is triggered by setting the environment variable
AGSDEBUG. it will cause emission of instruction numbers before
the assembly. it is not possible to recompile that output without
stripping the instruction numbers away.

+ some assertions to ensure we dont do invalid stuff with jmps.