(dis)assembler: detect pre-2.7.2 string type
the game 1213 episode 1 was made with 2.6.1 and uses a string type,
which uses a hardcoded length of 200.
the buffer is allocated in the data section, immediately followed
by a "pointer" to it (via FIXUP_DATADATA reference).
detecting this type removes a lot of spam and guesswork from the
variable list. unfortunately, the oldest AGS version available
is 2.7.2, which when using the type shows an error that now the
"String" type (uppercase S) has to be used.
the new type seems to be dynamically allocated.

we now detect the type keyword "string". to keep codechanges mini-
mal an assignment statement is emitted and required, however
ignored in the codegen.
note that it is possible the earlier engines allowed to have
such strings with pre-filled content. we only detect strings
that are all zero.