implement char array syntax
unreferenced and zero-set sequences of chars are now
joined into a single char array value with the following
syntax: char[512] varXXXX = 0.
this optimization removes a considerable amount of lines in
the .data section.
the game "tales of jayvin" has one gamescript (4) with a huge
array of 6 MB, and for every single byte a line was emitted,
which resulted in an .s file of 345 MB.

note that the array syntax is currently only implemented for
char, not short or int.