agsprite: restore alpha channel for 32bit img via "magenta"
this fixes the issues mentioned in
d3b5611d2dead118a0c05d67c0c4d731bf005717.

according to comment by @ivan-mogilko, information whether to use
alpha channel of 32bit image or magic magenta color is stored not
in the spritefile, but in other datafiles.

so in theory it's possible that there are some cases where game
authors used the alpha feature with images that didn't use magenta.
if such games are encountered, an option should be added to save
all 32bit images as such, and not discard alpha channel as is
currently done if image can be palettized (due to harddisk space
considerations).