agsprite: add hi-color-only-if-safe mode
this mode conservatively converts only 32bit sprites to 16bit where
it is entirely safe: this means only alpha values 0 and 0xff are
used, and transparency is only used for magenta color.

in verbose mode, id number of converted pics is printed so one
could grep the output for those filenumbers, then use agsalphahack
with a such created list to remove alpha flag bits in gamefile.

however, it's possible this is not necessary at all, testing this
with Quest For Infamy everything worked out of the box after conv-
ersion.

unfortunately, only a small number of files were converted, so
acsprset.spr size is 309 MB after conversion, 332MB without, i.e
<10% savings for this particular game.