disassembler: fix bug omitting exported vars not used in module
game `tales of jayvin` puts some variables in one script, then
uses them only from another, but not the one it was defined in.

repacking it with re-assembled scripts failed with import error.

when resolving variable offsets/names/sizes for data section
we therefore need to check for an export whenever fixup lookup
(means variable is used from .text) fails.

the reason export check is done after code check is that code
check can eventually figure out the correct type of the var,
if certain instructions are used on them, whereas export check
must assume the type is char for safety.