agssim: implement call/ret instructions
- fix label reference system
  the previous lazy way of simply adding negative offsets for
  unresolved labels was bogus.
  1) it wouldn't work with `li` instruction since we can't tell
     whether the parameter will be used for a later call insn
  2) it would only work with the first label reference since
     instruction address of the callsite is saved into bytecode.

- additionally improve some error messages

since the simulator starts executing code from IP 0, in order to
be able to declare functions without executing them in the order
they appear, a jump to a later defined label needs to be used,
e.g.:

    jmpi main
    function_return_1:
    li ax, 1
    ret
    main:
    li dx, function_return_1
    call dx

here the first instruction is jmpi main, so we can declare the
function "function_return_1" and jump over it, and use it only
when needed.

btw, as far as the simulator is concerned, labels and functions are
the same thing.