agssim: change syscall calling convention to "fastcall"
again, we do it (almost) like x86_64.
syscall number is passed in ax, args in bx, cx, dx.
if we add syscalls that need more than 3 args, eventually
op and mar could be used too.

having it this way makes it much easier (and shorter) to write
code doing syscalls.