agssim: fix bugs emitting strings with a size divisible by 4
the loop writes only l-1 bytes to keep the trailing '"' character
away, but intended to write (at least) one more nul byte to zero-
terminate the string. thus in case of l==0 the loop needs to run
another time. failure to do so causes a non-zero terminated string
and a mismatch between the size calculated for the jump and the
number of bytes actually written, which cause the jump to end up
in the middle of the string.