agssim: allocate 256 register slots and check accesses
the code used to access the array via the REGI macro via a
byte-code provided int index without any checks, causing a crash
due to another bug. since doing branches for every single register
access to check whether the value is valid or not is really costly,
and additionally can't be implemented as an inline func without
semantic change, as the macro is used for both rvalues and lvalues.

therefore the cheapest option was picked: 256 array slot and a
simple bit and with 0xff mask to keep the access within bounds
but without adding a performance penalty.