agssim: implement argv passing to program
command line options can now be passed to programs in the following
way:

    agssim -- options go here

or

    agssim file.s -- options go here

when program execution starts, register op contains a pointer to
an array of the following contents:
[argc]
[argv0]
...
[argvN]
[0]
[envvar0]
...
[envvar0]
[0]
[auxval0]
...
[auxvalN]
[0]

this is pretty much the same way x86_64 sysv abi does it.
the array contains pointer to strings, which are stored at the
beginning of the stack.
environment and aux are not yet implemented and always empty.