Clib32: preliminary support for new v30 datafile format
the file format was introduced to be able to deal with files > 2GB,
but we currently re-use the 32bit length/offset fields in the
multifile struct, so we can't deal yet with files/offsets > 4GB.
processing such a file will raise an assert error.
apparently the new format also removes some restrictions on number
of contained files and filename lengths, so this may also be something
to revisit.
also, the code to detect v30 format from file-end signature is currently
lacking.

this was implemented in upstream ags in ab7486dac6f4c6fb86c35df8f630b3926a7c540b