Script.c: fix segfault on encountering non-zero terminated string section
as seen in re-upload of "Leisure Suit Larry 2 Point and Click (Final version)"
( https://workupload.com/file/aT8tA93j ) in room55.
the version in the agsarchives.com torrent is broken too.

since the game is split into floppy-sized chunks using the multifile feature
of ags, it's pretty likely that the author had his copy on floppies and one
or more of them had some integrity issues. floppy disks were pretty unreliable
media and it was common to get CRC errors on rar or zip files saved on them.