agscriptxtract: initial support for recreating builtinscriptheader.ash
the header used internally by (older?) ags versions contains several
macros and enums filled in by objects defined in the editor GUI.

for example, when compiling the "default game" template in AGS 2.7.2,
the following header can be encountered in the process memory:

    #define AGS_MAX_CHARACTERS 300
    #define AGS_MAX_INV_ITEMS 301
    #define AGS_MAX_GUIS 50
    #define AGS_MAX_CONTROLS_PER_GUI 30
    #define AGS_MAX_VIEWS 600
    #define AGS_MAX_LOOPS_PER_VIEW 16
    #define AGS_MAX_FRAMES_PER_LOOP 20
    #define AGS_MAX_OBJECTS 20
    #define AGS_MAX_HOTSPOTS 30
    #define AGS_MAX_REGIONS 16
    enum CursorMode {
      eModeWalkto = 0,
      eModeLookat = 1,
      eModeInteract = 2,
      eModeTalkto = 3,
      eModeUseinv = 4,
      eModePickup = 5,
      eModePointer = 6,
      eModeWait = 7,
      eModeUsermode1 = 8,
      eModeUsermode2 = 9};
    import Character character[1];
    #define EGO 0
    import Character cEgo;
    import Hotspot hotspot[30];
    import Region region[16];
    import InventoryItem inventory[3];
    import InventoryItem iKey;
    import InventoryItem iPoster;
    import Dialog dialog[1];
    import GUI gui[3];
    #define STATUSLINE FindGUIID("STATUSLINE")
    import GUI gStatusline;
    #define ICONBAR FindGUIID("ICONBAR")
    import GUI gIconbar;
    import Button btnIconInv;
    import Button btnIconCurInv;
    import Button btnIconSave;
    import Button btnIconLoad;
    import Button btnIconExit;
    import Button btnIconAbout;
    #define INVENTORY FindGUIID("INVENTORY")
    import GUI gInventory;
    import InvWindow invCustomInv;
    import Button btnInvSelect;
    import Button btnInvLook;
    import Button btnInvOK;
    import Button btnInvDown;
    import Button btnInvUp;
    #define VIEW1 1
    #define VIEW2 2

the AGS_MAX constants seem to be hardcoded to the engine version's default.
the enum CursorMode contains all names used for mouse cursors, prefixed
with eMode and stripped of whitespace,

the import Character character[1]; contains the number of characters used
in the game as array bounds (but always using that name).

InventoryItem lines are added correspondingly to "Inventory items" gui tab,
the bounds of GUI and Dialog too.

the #define STATUSLINE FindGUIID("STATUSLINE") correspond to the names used
in the GUI tab, here the convention seems to be to always have the macro
name all uppercase, and the string as specified by the editor user.
additionally to the macro, a line like import GUI gStatusline; follows where
the convention seems to be a lowercase g followed by the user-chosen gui name.
this is the new OOP stuff introduced in AGS v 2.7.0. hail progress.

InvWindow invCustomInv; is a custom control of that name inside the INVENTORY
GUI, the Btn* imports too.

lastly, for each view in the view editor, the view name is defined as a macro
pointing to its index. this seems to be always all-uppercase according to CJ's
preferences.

this commit implements the basic framework to write the header, and starts with
exporting the cursor/mouse mode enum.