agscriptxtract: export dialog script sources, GUI header info
while wading through the huge amount of stuff processed by AGS
datafile format until it finally reaches the gui names, which
i was interested in, i noticed that older file formats save the
sourcecode for their dialog scripts in plain text, but encrypted.

i reverse-engineered the decryption function by creating
a dialog script in AGS 2.6.2 full of "a"s and quickly figured out
that it's just the reverse of the roomscript decryption code,
where characters of the string "Avis Durgan" need to be added to
the plaintext. in this case, they just need to be subtracted.
the code for this was even missing in the 2012 edition of the
AGS source code.

additionally i implemented what i originally had in mind, adding
the macros required to deal with GUIs in the generated header,
e.g:

    #define STATUSLINE FindGUIID("STATUSLINE")
    #define ICONBAR FindGUIID("ICONBAR")
    #define INVENTORY FindGUIID("INVENTORY")

in case of the AGS 2.6.2 default game template.
the only thing missing for a feature-complete header for AGS 2.6.2
support is viewnames.