make new datafile parsing code compatible with AGS 3.5.x
tested with "Last N' Furious" which was compiled with 3.4.x,
but it should work with 3.5.x era games too (hopefully).
especially the 3.5.x release has a ton of changes to the gui
code which requires special-casing.
it was a real pain in the ass digging through the code in AGS
and finding shortcircuits for all the added special cases.