agsprite: add support for reading spritecache version 10
introduced in ags commit 7077feacc90bd2771292608f0df0f6aa60aec16d,
the new 64 bit format seems to differ only in the handling of the
spriteindex file which has an array of 64bit offsets rather than
32 bits, but is otherwise identical, so it actually seems odd that
a new version number was assigned for the spritecache rather than
just for the sprite index. we never read a sprite index, but only
write them so no changes were necessary accept from writing our
spritecache in version 6 if the input version was higher than that,
and allowing the version 10 for reading.
unfortunately my assumptions from reading the commit couldn't be
tested yet due to lack of an AGS game using that spritecache version.