agsprite: add support for spritecache version 11
introduced in AGS commit 66bb2f7db86bf52f5533b0c44e474d9baa0ac27f, it
appears the only change is that num_sprites is read as a 32 bit integer
rather than 16 bit.

it's unclear why the commit message says "support for >32K sprites",
since an unsigned short can handle 64K values, apparently it was
either assumed the value is signed, or it is really signed, in which
case the obvious solution would have been to allow unsigned numbers
to start with (which would have doubled the amount of possible sprites)
and would probably have satisfied the game designers that hit the 32K
limit (unless there's special handling for signed values for errors,
in which case and could have changed the logic to use only a single value
for error, e.g. -1).
it's also possible the engine perfectly accepts an unsigned
value for the old format and the commit message was just wrong, this needs
to be tested and the MAX value added in this commit adapted accordingly.
using this change, the sprites of the game nocturna (noct191.exe) can be
successfully extracted and repacked, albeit the sprite numbers differ
after the repack. whether this is an actual issue needs to be evaluated.