agsprite: allow up to 64K-1 sprites to be packed into v6 format
as of AGS commit 4645eec145a9c095da3f2cd6b551ea3134526cf7, the
sprite number/count is processed as unsigned short, so we can make use
of the full 64K space for format v6.

0xffff aka (signed short)-1 is reserved for future use as "invalid"
value or similar for the engine, should that become necessary at
some point.

thanks go to @ivan-mogilko to make this possible and implementing
the change upstream.