minishilka: add experimental code for size-specific lookup
using the added lookup table that maps to the start index
for a specific length, we only need to iterate over the items
that have the same length as the passed keyword.

this slightly improves performance in my testcase which throws
a huge list of assembler keywords (73 possible keywords) from
4.69 seconds to 4.58, but it should have a much higher gain
if non-keywords are being passed to the find_keyword function.
however, this adds more complexity and probably makes performance
worse for usage with a small set (like 10) of keywords, because
in such a case its probably faster to just loop over 10 entries
instead of first doing a lot of operations to minimize the set
to look up, which introduces several branches.

in the end, this optimization is probably not worth it, but i
have some ideas to improve it still.