minishilka: improve speed and codesize with a return value table
a heuristics is used that tries to detect whether all keyword
actions meet certain simplicity criteria, if so, a table with
return values is generated instead of a huge switch statement.
this improves speed from about 4.0 secs to 3.85s on a huge input
file (in comparision, the real high-end shilka requires ~3.7s for
the same file).
more details are available in the usage/--help output.