add lex/yacc based parser for ags assembly
both were created about 1.5 years ago, but then abandoned for a while.
currently the parser works quite well, it supports the entire syntax
of .s files emitted by the disassembler. code is saved in a list of
basic blocks.
the parser currently has some embedded code that does some of the
assembly optimization pioneered by agsoptimize, but is much faster.
this code is meant to end up in a different TU, so the parser can be
used for a variety of purposes.
i'm checking this in now so there's finally a VCS snapshot available.