implement emission of section names in recompiled object files
the first section name contains the filename of the original source
file, and it is useful for debugging, so we no longer drop it.
dialogscripts contain more than one section name/offset pair, and
we currently emit only the first one because the offset (if used)
might become invalid after recompile and we'd have to add some
extra code to deal with it.