agspack: add support for games > 2GB
theoretically old MFL (multi-file-lib) formats should support files/offsets
up to 4GB just fine, however the AGS engine reads those fields as signed
integers, thereby limiting them to max 2GB in size.

the game "StrangeLand" is the first game i encountered that requires more
than that, in fact it has a single file "acsprset.spr" which is almost
2.9 GB.

in order to support files like those, we need to emit the MFL in version 30,
which is the first version to support 64 bit offsets and lengths.

note that the support for 64bit is still half-baked, in order to support files
or offsets > 4GB we would need to store them as 64bits in the structures
in Clib32.h, and then write them one-by-one (as is currently done for the
new v30 format writer), rather than as one single array-of-int writes.

closes #17