agsprite: fix bug packing sprites as 8 bit if they were optimized
as an optimization, sprites that have less than 256 colors are
stored as palettized 8 bit images in the targa format.
the code that decides whether the global palette* or the per-picture
palette was used had some logic involving spritecache version which
doesn't seem to make much sense, when looking at it today.
however, it's possible that spritecache version 5 "compressed" does
something special i forgot which isn't being considered after this
change.

*: images in ags that use 8bit all use a global palette that's stored
in the spritefile header.