agsprite: fix potential segfault extracting RLE from newer ags
newer AGS versions unpack the entire RLE-compressed buffer in a
single pass, unlike earlier versions that did it line-by-line.

this caused a crash when trying to do line-based uncompression
where the overflow check (if (n >= size) return 0; ) kicked in,
because the loop then assigned the 0 return value as the in ptr
and continued, dereferencing a null pointer.

whether it's safe to do "entire-file" rle uncompression also for
files created with older AGS versions remains to be seen, but
at least the AGS engine doesn't seem to have any checks for that
in place and presumably if it failed to load older spritesets,
someone would have reported it by now.