ADF_open: add fine grained error reporting, dont fail always
for some cases, we can ignore errors that happen during game file
parsing. for example, for sprite stuff we need only information
that's very early in the game file.
also for game scripts, most needed information comes relatively
early. the full parse is only needed to reproduce the header
with constants to recompile a game, but that's just a "nice-to-have"
and shouldnt prevent other functionality from kicking in when
the data can't be extracted.