agsinject: fix return value, rename on windows
the logic processing the return value from inject() was actually
wrong, it returned success in some cases even on error.
also, rename() on windows doesn't work like on unix, overwriting
the destination, therefore we need to delete the target first.
in my experiments with wine i couldn't get rename() to work anyway,
but this might be due to a bug in wine.
in such a case we now print the temp filename, so the user can
manually copy the file to its destination.

closes #29