add support for clib v30's hidden features
apart from using 64 bit offsets, v30 of clib also introduced
some new features, or rather, removed existing limitations:

1. number of files in a pack is no longer restricted to 10000
2. filenames may be up to 260 chars long (windows MAX_PATH)
3. filenames may contain subdirectories (using / as path delimiter)

since the old datastructs are completely expanded arrays for each
possible member at maximum length, implementing this required to
develop a new entirely dynamically-allocated container.
this new container is now used everywhere and uses a new highly-
optimized strstore component, written by yours truly.
with that, we can easily fix 1 and 2.
adding 3 to the mix was a bit of a nuisance, since it's not clear
whether there are limitations on the level of path components,
so i had to write code that iterately checks for and creates new
directories, and doing so in a safe manner that may not be used
by a rogue gamepack to overwrite /etc/shadow or your bashrc.
code was tested on roughly 30 ags titles from DOS era up to
the latest 2025 demo release of "old skies", which is the game
that leveraged all new features, containing a whopping 15k files
and several subdirs.

closes #38