agstract: only write fileno directives if more than 1 packfile
if there's only one packfile, the fileno is implicitly 0 for all
files. this makes the agspack.info much slimmer and easier to
navigate in most cases.

a new info file created this way doesn't break compatibility
with older agsutils binaries since the info was ignored anyway
(we always create a single big packfile), the fileno is only
stored in case we eventually see the need to support writing
multiple packfiles.