initial work to support games older than 2.50
while the exact layout of the game setup structs is still unknown,
we can at least search for the SCOM header in games back to 2.20,
and extract the scripts.
the existing code will still back out after the scripts are processed,
but they do get correctly extracted - including source code for
room scripts.
this might already prove useful for someone trying to forward-port
an old game.