agsprite: save 100% transparent 32bit images as 24bit
in blackwell legacy demo, sprites were encountered that have
all alpha bytes set to 0. the result is that such images appear
entirely transparent/invisible in image editing software such
as gimp.
it appears likely that the ags editor does this when the alphaflag
it stores separately for each picture, is unused.
this seems to be a bug, and it should set the alpha to 0xff
for all pixels in that case.
our fix is to simply convert these pictures to 24 bit instead
and complete leave the alpha bytes away.
that way the image can be properly edited.

on packing up, we check whether the original depth was 32bit
and the tga file is 24 bit, and restore the zero-alpha as it
originally was.

this solution allows us to not require the alphaflag array
from the main game file, which in turn requires another command
line parameter, and that the game data file exists to begin
with.

since the targa format specifies that 16 bit images have one
bit reserved for alpha on/off, which is incompatible with
AGS' 565 16 bit format, we always scale 16 bit images up to
24 bit too.
that means in the past if the pack process encountered a 24 bit
tga, it knew it was originally a 16 bit image. now it can also
be an originally 32bit image, but we can derive the original
bit depth from the filename.

closes #39