Makefile: using mingw, don't build asm preprocessor + other fixes
we also now use .obj as extension for windows object files,
so unix and coff object files can live side by side.

also add a fix that suppresses make's built-in rule for
extension-less files (executables on unix) that for some reason
are sometimes used instead of our %($EXE_EXT) pattern rule.
the result of that issue was that the compile would effectively do
`$CC agstract.c -o agstract`, completely ignoring the dependencies.
with that fixed, also a SECONDARY rule was added because it now
started to delete agstract.o directly after use.